+++ src/Faker/Internal/Types.hs
@@ -28,6 +28,7 @@ data SourceData
   | Computer
   | Control
   | DnD
+  | PearlJam
   | Departed
   | BossaNova
   | BreakingBad