+++ src/Faker/Music/PearlJam.hs
@@ -1,0 +1,18 @@
+{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE OverloadedStrings #-}
+
+module Faker.Music.PearlJam where
+
+import Data.Text (Text)
+import Faker (Fake(..))
+import Faker.Provider.PearlJam
+import Faker.TH
+
+$(generateFakeField "pearl_jam" "musicians")
+
+$(generateFakeField "pearl_jam" "albums")
+
+$(generateFakeField "pearl_jam" "songs")
+
+
+